Layed GroundWork for working Arcade Machines
When you enter the range of an arcade machine, you can press a button. When that happens it loads a new scene where you can play the game you were touching. I am planning on extending this, making the scene a smaller version in the scene you are already in and forcing the camera to look at the arcade machine, as if you are really playing the game. But for now, this at least gives me an idea of the next steps.